# openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "FIN-03 Agricultural Climate-Aware Credit Risk API",
    "version": "1.0.0",
    "description": "Proposed contract skeleton. Replace generic request/response placeholders with detailed schemas before implementation."
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/yield-predictions": {
      "post": {
        "operationId": "createYieldPrediction",
        "summary": "Create YieldPrediction",
        "tags": [
          "MVP"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/income-estimates": {
      "post": {
        "operationId": "createIncomeEstimate",
        "summary": "Create IncomeEstimate",
        "tags": [
          "MVP"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/assessments": {
      "post": {
        "operationId": "createAssessment",
        "summary": "Create Assessment",
        "tags": [
          "MVP"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/assessments/{assessment_id}": {
      "get": {
        "operationId": "getAssessment",
        "summary": "Get Assessment",
        "tags": [
          "MVP"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal error"
          }
        }
      },
      "parameters": [
        {
          "name": "assessment_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          },
          "description": "Unique assessment_id identifier"
        }
      ]
    },
    "/assessments/{assessment_id}/scenarios": {
      "post": {
        "operationId": "createScenario",
        "summary": "Create Scenario",
        "tags": [
          "MVP"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      },
      "get": {
        "operationId": "listScenarios",
        "summary": "List Scenarios",
        "tags": [
          "MVP"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal error"
          }
        }
      },
      "parameters": [
        {
          "name": "assessment_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          },
          "description": "Unique assessment_id identifier"
        }
      ]
    },
    "/assessments/{assessment_id}/explanations": {
      "get": {
        "operationId": "getAssessmentExplanations",
        "summary": "Get AssessmentExplanations",
        "tags": [
          "MVP"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal error"
          }
        }
      },
      "parameters": [
        {
          "name": "assessment_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          },
          "description": "Unique assessment_id identifier"
        }
      ]
    },
    "/borrowers/{borrower_id}/assessment-history": {
      "get": {
        "operationId": "getAssessmentHistory",
        "summary": "Get AssessmentHistory",
        "tags": [
          "MVP"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal error"
          }
        }
      },
      "parameters": [
        {
          "name": "borrower_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          },
          "description": "Unique borrower_id identifier"
        }
      ]
    },
...
    "/assessments/{assessment_id}/review-decision": {
      "post": {
        "operationId": "recordHumanReviewDecision",
        "summary": "Record HumanReviewDecision",
        "tags": [
          "Deferred"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation error"
          },
          "500": {
            "description": "Internal error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      },
      "parameters": [
        {
          "name": "assessment_id",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          },
          "description": "Unique assessment_id identifier"
        }
      ]
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}